requirements.txt
matplotlib
Pillow
numpy
scikit-learn
pytorch
torchvision
albumentations
streamlit